refactor: split DeclarationProcessPanel into two files
Extract VerticalStepper and related step components (StepCircle,
Step1/2/3Content, TransmittedRow, DeadlineRow) into a dedicated file.
Also remove unused `status` prop from Step2Content.

DeclarationProcessPanel.tsx: 146 lines
VerticalStepper.tsx: 282 lines
